refactor(exceptions): move to project root
