+++ terra/exceptions.py
